+++ src/main/kotlin/com/pylons/wallet/core/ops/getPylons.kt
@@ -7,8 +7,8 @@ import com.pylons.wallet.core.types.*
 
 internal fun getPylons(msg: MessageData): Response {
     require(msg.ints.containsKey(Keys.PYLONS)) { throw BadMessageException("getPylons", Keys.PYLONS, "Int") }
-    val tx = Core.getPylons(msg.ints[Keys.PYLONS]!!)
+    val tx = Core.getPylons(msg.ints[Keys.PYLONS]!!.toLong())
     waitUntilCommitted(tx.id!!)
     return Response(MessageData(strings = mutableMapOf(Keys.TX to tx.id!!)), Status.OK_TO_RETURN_TO_CLIENT)}
 
-fun Core.getPylons (q : Int) : Transaction = Core.engine.getPylons(q).submit()
+fun Core.getPylons (q : Long) : Transaction = Core.engine.getPylons(q).submit()
